Updates simc_worker to also create and store json_report

+++ server/app/workers/simc_worker.rb
@@ -4,16 +4,21 @@ class SimcWorker
 
   def perform(report_id)
     report = Simc::Report.find(report_id)
+    html_file_location = "#{Rails.root.join('reports')}/#{report.id}.html"
+    json_file_location = "#{Rails.root.join('reports')}/#{report.id}.json"
     stdout, stderr, status = Open3.capture3(
       "#{Rails.root.join('bin', 'simc')} armory=#{report.region},"\
-      "#{report.server},#{report.character} html=#{Rails.root.join('reports')}/#{report.id}.html"
+      "#{report.server},#{report.character} html=#{html_file_location} "\
+      "json2=#{json_file_location}"
     )
 
-    Discordrb::API::Channel.delete_message("Bot #{ENV['DISCORD_TOKEN']}", ENV['DISCORD_CHANNEL'], report.message_id)
+    Discordrb::API::Channel.delete_message("Bot #{ENV['DISCORD_TOKEN']}", ENV['DISCORD_CHANNEL'], report.message_id) if report.message_id
 
     if status == 0
-      report.update(html_report: File.read("#{Rails.root.join('reports')}/#{report.id}.html"))
-      File.delete("#{Rails.root.join('reports')}/#{report.id}.html")
+      report.update(html_report: File.read(html_file_location),
+                    json_report: File.read(json_file_location))
+      File.delete(html_file_location)
+      File.delete(json_file_location)
 
       puts "Generated simc report: #{report_id}"
       Discordrb::Bot.new(token: ENV['DISCORD_TOKEN']).send_message(
@@ -22,7 +27,7 @@ class SimcWorker
           Done simulating #{report.server}/#{report.character}
           View report at: #{ENV['APP_URL']}/simc/reports/#{report.id}
         MESSAGE
-        .tap { |string| string << "\nDEBUG: #{report.as_json(except: [:text_report, :html_report]).to_s}" if Rails.env.development? }
+        .tap { |string| string << "\nDEBUG: #{report.as_json(except: [:json_report, :html_report]).to_s}" if Rails.env.development? }
       )
     else
       Discordrb::Bot.new(token: ENV['DISCORD_TOKEN']).send_message(
@@ -31,7 +36,7 @@ class SimcWorker
           An error has occured while simming #{report.server}/#{report.character}:
           #{stderr}
         MESSAGE
-        .tap { |string| string << "\nDEBUG: #{report.as_json(except: [:text_report, :html_report]).to_s}" if Rails.env.development? }
+        .tap { |string| string << "\nDEBUG: #{report.as_json(except: [:json_report, :html_report]).to_s}" if Rails.env.development? }
       )
     end
   end
